Move cache to top-level
