+++ dask/diagnostics/__init__.py
@@ -1,3 +1,2 @@
 from .profile import Profiler
 from .progress import ProgressBar
-from .cache import Cache
